Ship exhaust, laser and sprite vertical wrap fixes
Change the ship exhaust graphic slightly to reduce the number of pixels
that stayed static from frame to frame to make it look a bit more
dynamic.

Sprites weren't wrapping around the top and bottom correctly and were
displaying out of range (intruding on attributes area).

The head or pulse part of the laser shot wasn't going right to the edge
of the screen (this was noticeable).