Baiter movement added, some graphic tweaks
Added the code to handle baiter movement.

Added a section to allow creating binary files for loading via
serial port or Spectranet.

Did some prep work for when the landscape will finally have to be scaled
to fit the Spectrum's screen area. The arcade game has about 250 pixels
to play with from top to bottom but Speccy only has 192. The landscape
looks obviously too high for the play area so this will have to be scaled
to about 75% of the original height.

Added some data which will be used by the particle engine (when written!)